step3 피드백 반영: view를 inputView, outputView로 분리
