+++ src/main/java/mission1/racing/RacingGame.java
@@ -1,11 +1,14 @@
 package mission1.racing;
 
+import mission1.racing.view.RacingGameInputView;
+import mission1.racing.view.RacingGameOutputView;
+
 public class RacingGame {
     private int time;
     private int carNumber;
 
     private Movement movement;
-    private RacingGameView view = new RacingGameView();
+    private RacingGameInputView inputView = new RacingGameInputView();
 
     public void startGame() {
         setGame();
@@ -16,8 +19,8 @@ public class RacingGame {
     }
 
     private void setGame() {
-        time = view.inputView("시도할 회수는 몇 회 인가요?");
-        carNumber = view.inputView("자동차 대수는 몇 대 인가요?");
+        time = inputView.inputView("시도할 회수는 몇 회 인가요?");
+        carNumber = inputView.inputView("자동차 대수는 몇 대 인가요?");
 
         movement = new Movement(carNumber);
     }
