+++ src/main/java/mission1/racing/view/RacingGameInputView.java
@@ -1,0 +1,13 @@
+package mission1.racing.view;
+
+import java.util.Scanner;
+
+public class RacingGameInputView {
+    private Scanner scanner = new Scanner(System.in);
+
+    public int inputView(String msg) {
+        System.out.println(msg);
+        return scanner.nextInt();
+    }
+
+}
